Refactor: split up query.js
